refactor: Move thread to thread package
